test(docx): consolidate code-block tests into test_backend_msword
Move code-block tests from new module to existing test_backend_msword.
Trim test_backend_msword to ≤1500 lines.
